requirements.txt
pyinstaller
gspread
python-dotenv
pytest